# This is a combination of 19 commits. # This is the 1st commit message:
Added Push to the dockerhub registry

# This is the commit message #2:

Modified context for dockerhub build

# This is the commit message #3:

Adding Array expensiend in the Prepare job.

Created an array of tags to push to the dockerhub registery

# This is the commit message #4:

Added an array test to github action

# This is the commit message #5:

Repo name was not formatted correctly

# This is the commit message #6:

Array of element

# This is the commit message #7:

From the documentation of the buildx docker.

The array of tags need to be in the array string format

# This is the commit message #8:

Transform tags as string

# This is the commit message #9:

Removed the option parameters

The default option is the value I use

# This is the commit message #10:

Modified join to the array

# This is the commit message #11:

Transform the Array into a StringArray

# This is the commit message #12:

Remove test for multiple tags

# This is the commit message #13:

Modified the dockerhub readme

# This is the commit message #14:

Addedd more section to the dockerhub readme

# This is the commit message #15:

Modified the Documentation

# This is the commit message #16:

Updated the Documentation for docker

# This is the commit message #17:

Added backward compatibility for the release branch

# This is the commit message #18:

Test if the tags is triggered

# This is the commit message #19:

Modified the PUSH and BUILD function