Added Push to the dockerhub registry
Modified context for dockerhub build

Adding Array expensiend in the Prepare job.

Created an array of tags to push to the dockerhub registery

Added an array test to github action

Repo name was not formatted correctly

Array of element

From the documentation of the buildx docker.

The array of tags need to be in the array string format

Transform tags as string

Removed the option parameters

The default option is the value I use

Modified join to the array

Transform the Array into a StringArray

Remove test for multiple tags

Modified the dockerhub readme

Addedd more section to the dockerhub readme

Modified the Documentation

Updated the Documentation for docker

Added backward compatibility for the release branch

Test if the tags is triggered

Modified the PUSH and BUILD function

chore(dev): reorganize repository (#5153)

Added Push to the dockerhub registry

Modified context for dockerhub build

Adding Array expensiend in the Prepare job.

Created an array of tags to push to the dockerhub registery

Added an array test to github action

Repo name was not formatted correctly

Array of element

From the documentation of the buildx docker.

The array of tags need to be in the array string format

Transform tags as string

Removed the option parameters

The default option is the value I use

Modified join to the array

Transform the Array into a StringArray

Remove test for multiple tags

Modified the dockerhub readme

Addedd more section to the dockerhub readme

Modified the Documentation

Updated the Documentation for docker

Added backward compatibility for the release branch

Test if the tags is triggered

Modified the PUSH and BUILD function

chore(dev): reorganize repository (#5153)