Update package.json
The error message suggests that there is an issue with fetching the node-svm package from the yarnpkg registry. The request to fetch the package returned a "404 Not Found" error, which means that the package was not found in the registry.

To fix this issue, updating the version of the node-svm package in  package.json file to a version that exists in the yarnpkg registry. Also can  try running yarn install to ensure that all dependencies are up-to-date and installed correctly.