Pass through payload value for WhatsApp button messages
When sending [Message Templates](https://developers.facebook.com/docs/whatsapp/cloud-api/guides/send-message-templates#interactive),
the WhatsApp Cloud API allows the developer to attach a custom `payload` string to each of the quick
reply action buttons, which will be sent along with the text to the webhook callback.

I can see the payload in the log being emitted on line 42 in this file:

```
[whatsapp] Received button message from Whatsapp: { payload: 'PAYLOAD_1', text: 'My Button Text' }
```

But there's currently no way to access it from the bot workflow. Passing
it through here would really help with maintaining context during a
bot-initiated whatsapp conversation.