fix(integrations/whatsapp): flaky repeat test (#14270)
Co-authored-by: François Levasseur <francois_levasseur@hotmail.com>

After investigating, I found that the code looks to work exactly as intended. However, it appears that the "setTimeout" call in the sleep function will occasionally call it's callback handler a few fractions of a millisecond earlier than it should. This in turn causes a rounding issue in the Date.now() used to collect the timestamps which will then produce a delta that is ~1ms below the expected result in the test assertions.

As a result, it was decided that adding a small buffer in the assertions would be the simplest fix to the issue.